move pathing to separate module
